refactor: 프리젠테이션 레이어에서 User 도메인 사용
1. UserSessionInfo가 User를 필드로 가짐
2. HttpSession "user" 애트리뷰트에 User 객체 보관
3. Login 요청시에는 UserDto 대신  LoginDto 이용
4. 도메인 메서드 인자 변경 String -> Entity
5. 도메인 setter 제거

+++ src/main/java/techcourse/myblog/service/ArticleService.java
@@ -6,12 +6,15 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import techcourse.myblog.domain.article.Article;
 import techcourse.myblog.domain.article.ArticleException;
+import techcourse.myblog.domain.user.User;
+import techcourse.myblog.domain.user.UserException;
 import techcourse.myblog.dto.ArticleDto;
 import techcourse.myblog.repository.ArticleRepository;
 
 import java.util.List;
 
 @Service
+@Transactional
 public class ArticleService {
     private static final String NOT_EXIST_ARTICLE = "해당 아티클이 없습니다.";
     private final ArticleRepository articleRepository;
@@ -36,21 +39,23 @@ public class ArticleService {
         return articleRepository.findById(articleId).orElseThrow(() -> new ArticleException(NOT_EXIST_ARTICLE));
     }
 
-    @Transactional
     public Long save(Article article) {
         Article newArticle = articleRepository.save(article);
         return newArticle.getId();
     }
 
-    @Transactional
-    public Article update(long articleId, ArticleDto articleDto) {
+    public Article update(long articleId, ArticleDto articleDto, User author) {
         Article originArticle = findArticle(articleId);
-        originArticle.update(articleDto.toEntity());
+        originArticle.update(articleDto.toEntity(author));
         return originArticle;
     }
 
-    @Transactional
-    public void delete(long articleId) {
-        articleRepository.deleteById(articleId);
+    public void delete(long articleId, User author) {
+        Article article = findArticle(articleId);
+        if (article.isAuthor(author)) {
+            articleRepository.deleteById(articleId);
+            return;
+        }
+        throw new UserException();
     }
 }
